chore: update redirection logic and add hooks for checking hub redirection
